+++ src/cc-context/index.js
@@ -22,9 +22,6 @@ const ccContext = {
   moduleName_globalStateKeys_: {
 
   },
-  moduleName_sharedKeysWhichMapToGlobal_: {
-
-  },
   //to let cc know which ccClass are watching globalStateKeys
   globalCcClassKeys: [],
   /**
@@ -38,8 +35,20 @@ const ccContext = {
   ccClassKey_ccClassContext_: {
 
   },
-  // [globalKey]:{module:'xxx',key:'yyy'}
+   // [globalKey]:${modules}, let cc know what modules are watching a same globalKey
+  globalKey_toModules_: {
+
+  },
+  // [globalKey]:${sharedKey}
   globalMappingKey_sharedKey_: {
+
+  },
+  // [globalKey]:${modules}, let cc know what modules are watching a same globalMappingKey
+  globalMappingKey_toModules_: {
+
+  },
+  // let cc know a globalMappingKey is mapped from which module
+  globalMappingKey_fromModule_: {
 
   },
   // globalStateKeys is maintained by cc automatically,
@@ -50,7 +59,7 @@ const ccContext = {
 
   ],
   //  all global keys that exclude sharedToGlobalMapping keys
-  pureGlobalStateKeys:[
+  pureGlobalStateKeys: [
 
   ],
   store: {