+++ src/core/helper/map-shared-key-to-global.js
@@ -1,11 +1,13 @@
 import ccContext from '../../cc-context';
 import { MODULE_GLOBAL } from '../../support/constant';
+import util from '../../support/util';
 
 export default function (moduleName, sharedKey, globalMappingKey) {
   const _state = ccContext.store._state;
   const globalMappingKey_sharedKey_ = ccContext.globalMappingKey_sharedKey_;
+  const globalMappingKey_toModules_ = ccContext.globalMappingKey_toModules_;
+  const globalMappingKey_fromModule_ = ccContext.globalMappingKey_fromModule_;
   const globalStateKeys = ccContext.globalStateKeys;
-  const moduleName_sharedKeysWhichMapToGlobal_ = ccContext.moduleName_sharedKeysWhichMapToGlobal_;
   const globalState = _state[MODULE_GLOBAL];
   const moduleState = _state[moduleName];
 
@@ -18,11 +20,9 @@ export default function (moduleName, sharedKey, globalMappingKey) {
 
   globalStateKeys.push(globalMappingKey);
   globalState[globalMappingKey] = moduleState[sharedKey];
-  globalMappingKey_sharedKey_[globalMappingKey] = { module: moduleName, key: sharedKey };
+  globalMappingKey_sharedKey_[globalMappingKey] = sharedKey;
+  globalMappingKey_fromModule_[globalMappingKey] = moduleName;
 
-  let sharedKeysWhichMapToGlobal = moduleName_sharedKeysWhichMapToGlobal_[moduleName];
-  if(!sharedKeysWhichMapToGlobal){
-    sharedKeysWhichMapToGlobal = moduleName_sharedKeysWhichMapToGlobal_[moduleName] = [];
-  }
-  sharedKeysWhichMapToGlobal.push(sharedKey);
+  const mappingKeyModules = util.safeGetArrayFromObject(globalMappingKey_toModules_, globalMappingKey);
+  mappingKeyModules.push(moduleName);
 }